google_cloud_function/requirements.txt
# functions-framework==3.*
bs4
requests
google.cloud
google-cloud-storage